Remove any code requiring string resource form ExoPlayerManager. Move dependency of ErrorMessageProvider out of SimpleExoPlayerManager;
